+++ src/config/tool-assets.ts
@@ -1,2 +1,3 @@
 export const QUICK_SHIP_IMAGE_URL = 'https://lzk-web.oss-cn-beijing.aliyuncs.com/img/share/quick-ship.png'
 export const RELAY_ICON_URL = 'https://lzk-web.oss-cn-beijing.aliyuncs.com/img/share/relay-icon.jpg'
+export const SWC_LOGO_IMAGE_URL = 'https://lzk-web.oss-cn-beijing.aliyuncs.com/img/share/swc-logo.png'